完成商家菜品及菜品分类模块

+++ shop.blem.com/app/Http/Controllers/MenuCategoriesController.php
@@ -2,12 +2,18 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Menu;
 use App\Models\MenuCategories;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class MenuCategoriesController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -15,10 +21,10 @@ class MenuCategoriesController extends Controller
      */
     public function index(Request $request)
     {
-        $rows = MenuCategories::paginate(1);//获取分页数据
+        $rows = MenuCategories::where('shop_id',Auth::user()->id)->paginate(6);//获取分页数据
 
         if($keyword = $request->keyword){
-            $rows = MenuCategories::where('name','like',"%{$keyword}%")->paginate(1);//获取查询分页数据
+            $rows = MenuCategories::where('name','like',"%{$keyword}%")->paginate(6);//获取查询分页数据
         }
         //菜品分类列表
         return view('Menus.index',compact('rows','keyword'));
@@ -48,20 +54,21 @@ class MenuCategoriesController extends Controller
         $this->validate($request,
             [
                 'name' => 'required',//名称
-                'type_accumulation' => 'required|alpha',//菜品编号（a-z前端使用）
                 'description' => 'required',//描述
             ],
             [
                 'name.required' => '分类名称不能为空',
-                'type_accumulation.required' => '菜品编号不能为空',
-                'type_accumulation.alpha' => '菜品编号只能是字母',
                 'description.required' => '描述不能为空',
             ]);
 
+        //查询当前所有分类的默认情况
+        if($request->is_selected){
+            $this->cate();//将所有分类改为0
+        }
         //验证通过
         MenuCategories::create([
            'name' => $request->name,
-           'type_accumulation' => $request->type_accumulation,
+           'type_accumulation' => $this->strShuffle(),
             'description' => $request->description,
             'shop_id' => Auth::user()->id,
             'is_selected' => $request->is_selected,
@@ -107,20 +114,21 @@ class MenuCategoriesController extends Controller
         $this->validate($request,
             [
                 'name' => 'required',//名称
-                'type_accumulation' => 'required|alpha',//菜品编号（a-z前端使用）
                 'description' => 'required',//描述
             ],
             [
                 'name.required' => '分类名称不能为空',
-                'type_accumulation.required' => '菜品编号不能为空',
-                'type_accumulation.alpha' => '菜品编号只能是字母',
                 'description.required' => '描述不能为空',
             ]);
+
+        //查询当前所有分类的默认情况
+        if($request->is_selected){
+            $this->cate();//将所有分类改为0
+        }
 
         //验证通过
         $menu->update([
             'name' => $request->name,
-            'type_accumulation' => $request->type_accumulation,
             'description' => $request->description,
             'is_selected' => $request->is_selected,
         ]);
@@ -137,9 +145,26 @@ class MenuCategoriesController extends Controller
      */
     public function destroy(MenuCategories $menu)
     {
+        if( Menu::where('category_id',$menu->id)->first() ){
+            return back()->with('danger','该分类下还拥有菜品，不能进行删除！');
+        }
         //删除
         $menu->delete();
 
-        return redirect()->route('menus.index')->with('success','删除菜品分类成功！');
+        return redirect()->back()->with('success','删除菜品分类成功！');
+    }
+
+
+
+    //只能有一个默认分类(将所有分类的默认修改为否)
+    protected function cate(){
+       DB::update('update menu_categories set is_selected = 0 where shop_id = ?',[Auth::user()->id]);
+    }
+
+    //更好随机5位字符
+    protected function strShuffle(){
+        $str = 'qwertyuiopasdfghjklzxcvbnm';//字符集合
+        $str = str_shuffle($str);//随机打乱
+        return str_random(5,$str);//返回前5位
     }
 }
